Add authentication

+++ src/main/java/com/example/autoshopserver/car/CarService.java
@@ -1,12 +1,15 @@
 package com.example.autoshopserver.car;
 
+import com.example.autoshopserver.auth.info.AuthInfoService;
 import com.example.autoshopserver.car.brand.*;
-import com.example.autoshopserver.exception.exceptions.BrandNotFoundException;
 import com.example.autoshopserver.exception.exceptions.CarNotFoundException;
 import com.example.autoshopserver.exception.exceptions.UserNotFoundException;
+import com.example.autoshopserver.user.Role;
 import com.example.autoshopserver.user.User;
 import com.example.autoshopserver.user.UserRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -20,47 +23,44 @@ public class CarService {
     private final BrandService brandService;
     private final UserRepository userRepository;
     private final BrandMapper brandMapper;
+    private final AuthInfoService authInfoService;
 
-    public List<CarDTO> getCarList() {
+    public List<CarDTO> getCarList(Authentication authentication) {
         return carMapper.toCarDTOs(carRepository.findAll());
     }
 
-    public CarDTO getCarById(Long id) {
+    public CarDTO getCarById(Long id, Authentication authentication) {
         Car car = carRepository.findById(id).orElseThrow(CarNotFoundException::new);
         return carMapper.toCarDTO(car);
     }
 
-    public CarDTO saveCar(Long id, CarDTO carDTO) {
+    public CarDTO saveCar(Long id, CarDTO carDTO, Authentication authentication) {
         Car car = carMapper.toCar(carDTO);
-        if (id != null) car.setId(id);
-        BrandDTO brandDTO = brandService.getBrandById(carDTO.getBrandId());
+        if (id != null) {
+            User currentUser = authInfoService.getUserByAuthentication(authentication);
+            if (!car.getOwner().getId().equals(currentUser.getId()) && !currentUser.getRoles().contains(Role.ADMIN)) {
+                throw new AccessDeniedException("The advertisement does not belong to the user");
+            }
+            car.setId(id);
+            car.setOwner(authInfoService.getUserByAuthentication(authentication));
+        }
+        else {
+            User owner = userRepository.findById(carDTO.getOwnerId()).orElseThrow(UserNotFoundException::new);
+            car.setOwner(owner);
+        }
+        BrandDTO brandDTO = brandService.getBrandById(carDTO.getBrandId(), authentication);
         car.setBrand(brandMapper.toBrand(brandDTO));
         //User owner = userRepository.findById(carDTO.getOwnerId()).orElseThrow(UserNotFoundException::new);
-        User owner = userRepository.findById(2L).orElseThrow(UserNotFoundException::new);
-        car.setOwner(owner);
         return carMapper.toCarDTO(carRepository.save(car));
     }
+    
 
-    /*public CarDTO createCar(CarDTO carDTO) {
-        Car car = carMapper.toCar(carDTO);
-        BrandDTO brandDTO = brandService.getBrandById(carDTO.getBrandId());
-        car.setBrand(brandMapper.toBrand(brandDTO));
-        User owner = userRepository.findById(carDTO.getOwnerId()).orElseThrow(UserNotFoundException::new);
-        car.setOwner(owner);
-        return carMapper.toCarDTO(carRepository.save(car));
-    }
-
-    public CarDTO updateCar(Long id, CarDTO carDTO) {
-        Car car = carMapper.toCar(carDTO);
-        car.setId(id);
-        BrandDTO brandDTO = brandService.getBrandById(carDTO.getBrandId());
-        car.setBrand(brandMapper.toBrand(brandDTO));
-        User owner = userRepository.findById(carDTO.getOwnerId()).orElseThrow(UserNotFoundException::new);
-        car.setOwner(owner);
-        return carMapper.toCarDTO(carRepository.save(car));
-    }*/
-
-    public void deleteCar(Long carId) {
+    public void deleteCar(Long carId, Authentication authentication) {
+        Car car = carRepository.findById(carId).orElseThrow(CarNotFoundException::new);
+        User currentUser = authInfoService.getUserByAuthentication(authentication);
+        if (!car.getOwner().getId().equals(currentUser.getId()) && !currentUser.getRoles().contains(Role.ADMIN)) {
+            throw new AccessDeniedException("The advertisement does not belong to the user");
+        }
         carRepository.deleteById(carId);
     }
